--- a/_compat.c
+++ b/_compat.c
@@ -6,7 +6,8 @@ int main() {
     for (int i = 0; i < 4096; i++) {
         buf[i] = (unsigned char)((i+1)%251);
 
-        uint64_t h = XXH3_64bits(buf, (size_t)i);
+        uint64_t s = XXH3_64bits(buf, (size_t)i);
+        uint64_t h = XXH3_64bits_withSeed(buf, (size_t)i, s);
         printf("\t0x%llx, ", h);
 
         if (i % 4 == 3) {