--- a/utils.go
+++ b/utils.go
@@ -22,8 +22,10 @@ func readU16(p ptr, o ui) uint16 { return le.Uint16((*[2]byte)(ptr(ui(p) + o))[:
 func readU32(p ptr, o ui) uint32 { return le.Uint32((*[4]byte)(ptr(ui(p) + o))[:]) }
 func readU64(p ptr, o ui) uint64 { return le.Uint64((*[8]byte)(ptr(ui(p) + o))[:]) }
 
+func writeU64(p ptr, o ui, v u64) { le.PutUint64((*[8]byte)(ptr(ui(p) + o))[:], v)}
+
 func xxhAvalancheSmall(x u64) u64 {
-	x ^= u64(key32_000 ^ key32_004)
+	x ^= x >> 33
 	x *= prime64_2
 	x ^= x >> 29
 	x *= prime64_3