--- a/vector_hash_other.go
+++ b/vector_hash_other.go
@@ -7,4 +7,4 @@ const (
 	sse2 = false
 )
 
-func hashVector(_ ptr, _ u64) u64 { return 0 }
+func hashVector(_ ptr, _ u64, _ ptr) u64 { return 0 }