As Cyan mentioned in his [blog](https://fastcompression.blogspot.com/2019/03/presenting-xxh3.html), xxh3 has initially developed for bloom filters that need changing seed for each hashing. Now I make a bloom filter in Go in my project so I have written this patch.

However, the performance gets slower due to:
1. extra instructions to add/sub keys
2. disarm of some immediate value hacks
3. larger stack and copying key to an array in case of hashing large bytes.

I'm making this PR just for saying thanks for your work. It's another option to host separate codes both high performance and seed support.

Here are benchmarks on my Intel Core i7-4790K CPU @ 4.00GHz.

master vs seed = 0
```
$benchstat master seed0
name                   master time/op  seed0 time/op  delta
Fixed/0-8              3.58ns ± 8%    5.70ns ± 6%  +59.20%  (p=0.000 n=95+92)
Fixed/1-8              4.03ns ± 4%    5.56ns ± 5%  +38.12%  (p=0.000 n=93+94)
Fixed/2-8              4.04ns ± 4%    5.62ns ± 6%  +38.87%  (p=0.000 n=95+96)
Fixed/3-8              3.69ns ± 3%    5.50ns ± 6%  +49.34%  (p=0.000 n=92+91)
Fixed/4-8              4.58ns ± 4%    6.36ns ± 5%  +38.71%  (p=0.000 n=91+91)
Fixed/8-8              4.58ns ± 4%    6.37ns ± 6%  +38.95%  (p=0.000 n=97+92)
Fixed/9-8              4.38ns ± 5%    6.00ns ± 7%  +37.12%  (p=0.000 n=95+92)
Fixed/16-8             4.38ns ± 4%    6.01ns ± 6%  +37.26%  (p=0.000 n=98+95)
Fixed/17-8             7.04ns ± 4%    8.32ns ± 5%  +18.09%  (p=0.000 n=95+94)
Fixed/32-8             7.03ns ± 4%    8.28ns ± 5%  +17.68%  (p=0.000 n=94+88)
Fixed/33-8             10.0ns ± 7%    11.3ns ± 6%  +13.24%  (p=0.000 n=97+94)
Fixed/64-8             10.0ns ± 4%    11.4ns ± 7%  +14.09%  (p=0.000 n=96+93)
Fixed/65-8             12.9ns ± 5%    14.3ns ± 6%  +10.96%  (p=0.000 n=96+97)
Fixed/96-8             12.9ns ± 4%    14.3ns ± 7%  +11.12%  (p=0.000 n=91+93)
Fixed/97-8             15.5ns ± 4%    16.9ns ± 6%   +8.57%  (p=0.000 n=95+92)
Fixed/128-8            15.5ns ± 4%    16.9ns ± 5%   +8.58%  (p=0.000 n=92+90)
Fixed/129-8            18.0ns ± 4%    20.2ns ± 5%  +11.97%  (p=0.000 n=95+93)
Fixed/240-8            34.1ns ± 4%    39.1ns ± 6%  +14.72%  (p=0.000 n=97+93)
Fixed/241-AVX2-8       65.9ns ± 4%    68.5ns ± 4%   +3.91%  (p=0.000 n=97+94)
Fixed/241-SSE2-8       23.7ns ± 4%    27.4ns ± 5%  +15.61%  (p=0.000 n=93+95)
Fixed/241-8            42.1ns ± 3%    44.9ns ± 5%   +6.78%  (p=0.000 n=94+93)
Fixed/512-AVX2-8       70.2ns ± 4%    73.8ns ± 4%   +5.12%  (p=0.000 n=96+93)
Fixed/512-SSE2-8       33.4ns ± 4%    37.2ns ± 4%  +11.18%  (p=0.000 n=94+94)
Fixed/512-8            73.5ns ± 4%    77.2ns ± 6%   +5.10%  (p=0.000 n=93+94)
Fixed/1024-AVX2-8      82.3ns ± 4%    86.3ns ± 5%   +4.85%  (p=0.000 n=99+95)
Fixed/1024-SSE2-8      53.5ns ± 5%    57.1ns ± 4%   +6.84%  (p=0.000 n=97+93)
Fixed/1024-8            135ns ± 3%     141ns ± 4%   +4.61%  (p=0.000 n=92+93)
Fixed/8192-AVX2-8       235ns ± 5%     243ns ± 4%   +3.19%  (p=0.000 n=95+95)
Fixed/8192-SSE2-8       353ns ± 4%     355ns ± 5%     ~     (p=0.329 n=92+93)
Fixed/8192-8           1.06µs ± 5%    1.10µs ± 7%   +3.82%  (p=0.000 n=96+96)
Fixed/102400-AVX2-8    2.32µs ± 4%    2.39µs ± 4%   +3.06%  (p=0.000 n=92+92)
Fixed/102400-SSE2-8    4.31µs ± 4%    4.39µs ± 6%   +1.81%  (p=0.000 n=90+96)
Fixed/102400-8         13.2µs ± 3%    13.7µs ± 5%   +3.91%  (p=0.000 n=92+98)
```

master vs seed = 42
```
$benchstat master seed42
name                   master time/op seed42 time/op delta
Fixed/0-8              3.58ns ± 8%    5.76ns ± 6%   +60.96%  (p=0.000 n=95+47)
Fixed/1-8              4.03ns ± 4%    5.83ns ± 7%   +44.78%  (p=0.000 n=93+44)
Fixed/2-8              4.04ns ± 4%    5.90ns ±11%   +45.87%  (p=0.000 n=95+47)
Fixed/3-8              3.69ns ± 3%    5.72ns ± 7%   +55.28%  (p=0.000 n=92+48)
Fixed/4-8              4.58ns ± 4%    6.58ns ± 6%   +43.50%  (p=0.000 n=91+45)
Fixed/8-8              4.58ns ± 4%    6.59ns ± 4%   +43.77%  (p=0.000 n=97+43)
Fixed/9-8              4.38ns ± 5%    6.16ns ± 6%   +40.72%  (p=0.000 n=95+48)
Fixed/16-8             4.38ns ± 4%    6.20ns ± 6%   +41.44%  (p=0.000 n=98+47)
Fixed/17-8             7.04ns ± 4%    8.11ns ± 3%   +15.22%  (p=0.000 n=95+42)
Fixed/32-8             7.03ns ± 4%    8.18ns ± 5%   +16.30%  (p=0.000 n=94+48)
Fixed/33-8             10.0ns ± 7%    11.2ns ± 5%   +11.83%  (p=0.000 n=97+45)
Fixed/64-8             10.0ns ± 4%    11.2ns ± 4%   +12.02%  (p=0.000 n=96+45)
Fixed/65-8             12.9ns ± 5%    14.1ns ± 7%    +9.03%  (p=0.000 n=96+45)
Fixed/96-8             12.9ns ± 4%    14.1ns ± 7%    +9.39%  (p=0.000 n=91+44)
Fixed/97-8             15.5ns ± 4%    16.7ns ± 4%    +7.50%  (p=0.000 n=95+48)
Fixed/128-8            15.5ns ± 4%    16.8ns ± 8%    +8.09%  (p=0.000 n=92+48)
Fixed/129-8            18.0ns ± 4%    20.2ns ± 5%   +11.90%  (p=0.000 n=95+47)
Fixed/240-8            34.1ns ± 4%    39.0ns ± 3%   +14.53%  (p=0.000 n=97+43)
Fixed/241-AVX2-8       65.9ns ± 4%    99.7ns ± 5%   +51.35%  (p=0.000 n=97+46)
Fixed/241-SSE2-8       23.7ns ± 4%    54.7ns ± 7%  +131.21%  (p=0.000 n=93+45)
Fixed/241-8            42.1ns ± 3%    72.0ns ± 4%   +71.29%  (p=0.000 n=94+46)
Fixed/512-AVX2-8       70.2ns ± 4%   104.7ns ± 4%   +49.00%  (p=0.000 n=96+47)
Fixed/512-SSE2-8       33.4ns ± 4%    63.9ns ± 6%   +91.06%  (p=0.000 n=94+45)
Fixed/512-8            73.5ns ± 4%   103.9ns ± 8%   +41.40%  (p=0.000 n=93+46)
Fixed/1024-AVX2-8      82.3ns ± 4%   117.1ns ± 3%   +42.27%  (p=0.000 n=99+45)
Fixed/1024-SSE2-8      53.5ns ± 5%    83.5ns ± 4%   +56.21%  (p=0.000 n=97+45)
Fixed/1024-8            135ns ± 3%     169ns ± 7%   +25.03%  (p=0.000 n=92+44)
Fixed/8192-AVX2-8       235ns ± 5%     276ns ± 5%   +17.43%  (p=0.000 n=95+46)
Fixed/8192-SSE2-8       353ns ± 4%     419ns ± 8%   +18.60%  (p=0.000 n=92+45)
Fixed/8192-8           1.06µs ± 5%    1.14µs ± 6%    +7.57%  (p=0.000 n=96+47)
Fixed/102400-AVX2-8    2.32µs ± 4%    2.44µs ± 6%    +5.25%  (p=0.000 n=92+46)
Fixed/102400-SSE2-8    4.31µs ± 4%    4.56µs ± 8%    +5.88%  (p=0.000 n=90+48)
Fixed/102400-8         13.2µs ± 3%    13.8µs ± 5%    +4.74%  (p=0.000 n=92+46)
```