@klauspost 2021-09-16
@peroxyacyl I've redone the PR in #12 - duplicated the code and added it to the Hasher.

There is no 128 bit hash support yet. Do you know where I can find the information to do `hashSmall128` and `hashMed128`?