@zeebo 2021-11-11
Seed support has been added and in v1.0.0 now.